opengl

+++ include/WZE/camera.hpp
@@ -6,25 +6,25 @@
 namespace wze {
     class camera final {
         private:
-        static float _x;
-        static float _y;
-        static float _z;
-        static float _angle;
-        static float _fov;
-        static float _aspect;
+        static double _x;
+        static double _y;
+        static double _z;
+        static double _angle;
+        static double _fov;
+        static double _aspect;
         static void _update(void);
 
         public:
-        static auto x(void) -> float;
-        static void set_x(float x);
-        static auto y(void) -> float;
-        static void set_y(float y);
-        static auto z(void) -> float;
-        static void set_z(float z);
-        static auto angle(void) -> float;
-        static void set_angle(float angle);
-        static auto fov(void) -> float;
-        static void set_fov(float fov);
-        static void __init(float aspect);
+        static auto x(void) -> double;
+        static void set_x(double x);
+        static auto y(void) -> double;
+        static void set_y(double y);
+        static auto z(void) -> double;
+        static void set_z(double z);
+        static auto angle(void) -> double;
+        static void set_angle(double angle);
+        static auto fov(void) -> double;
+        static void set_fov(double fov);
+        static void __init(double aspect);
     };
 } // namespace wze
